update the usage of factcheck as a library
